REF: share maybe_cast_slice_bound DTI/TDI/PI (#42897)
