+++ pandas/core/indexes/datetimes.py
@@ -25,7 +25,6 @@ from pandas._libs import (
 )
 from pandas._libs.tslibs import (
     Resolution,
-    parsing,
     timezones,
     to_offset,
 )
@@ -686,48 +685,10 @@ class DatetimeIndex(DatetimeTimedeltaMixin):
             key = key.tz_convert(self.tz)
         return key
 
+    @doc(DatetimeTimedeltaMixin._maybe_cast_slice_bound)
     def _maybe_cast_slice_bound(self, label, side: str, kind=lib.no_default):
-        """
-        If label is a string, cast it to datetime according to resolution.
-
-        Parameters
-        ----------
-        label : object
-        side : {'left', 'right'}
-        kind : {'loc', 'getitem'} or None
-
-        Returns
-        -------
-        label : object
-
-        Notes
-        -----
-        Value of `side` parameter should be validated in caller.
-        """
-        assert kind in ["loc", "getitem", None, lib.no_default]
-        self._deprecated_arg(kind, "kind", "_maybe_cast_slice_bound")
-
-        if isinstance(label, str):
-            try:
-                parsed, reso = self._parse_with_reso(label)
-            except parsing.DateParseError as err:
-                raise self._invalid_indexer("slice", label) from err
-
-            lower, upper = self._parsed_string_to_bounds(reso, parsed)
-            # lower, upper form the half-open interval:
-            #   [parsed, parsed + 1 freq)
-            # because label may be passed to searchsorted
-            # the bounds need swapped if index is reverse sorted and has a
-            # length > 1 (is_monotonic_decreasing gives True for empty
-            # and length 1 index)
-            if self._is_strictly_monotonic_decreasing and len(self) > 1:
-                return upper if side == "left" else lower
-            return lower if side == "left" else upper
-        elif isinstance(label, self._data._recognized_scalars):
-            self._deprecate_mismatched_indexing(label)
-        else:
-            raise self._invalid_indexer("slice", label)
-
+        label = super()._maybe_cast_slice_bound(label, side, kind=kind)
+        self._deprecate_mismatched_indexing(label)
         return self._maybe_cast_for_get_loc(label)
 
     def slice_indexer(self, start=None, end=None, step=None, kind=lib.no_default):
@@ -804,7 +765,7 @@ class DatetimeIndex(DatetimeTimedeltaMixin):
             return indexer
 
     @doc(Index.get_slice_bound)
-    def get_slice_bound(self, label, side: str, kind=None) -> int:
+    def get_slice_bound(self, label, side: str, kind=lib.no_default) -> int:
         # GH#42855 handle date here instead of _maybe_cast_slice_bound
         if isinstance(label, date) and not isinstance(label, datetime):
             label = Timestamp(label).to_pydatetime()
