+++ pandas/core/indexes/timedeltas.py
@@ -178,37 +178,6 @@ class TimedeltaIndex(DatetimeTimedeltaMixin):
 
         return Index.get_loc(self, key, method, tolerance)
 
-    def _maybe_cast_slice_bound(self, label, side: str, kind=lib.no_default):
-        """
-        If label is a string, cast it to timedelta according to resolution.
-
-        Parameters
-        ----------
-        label : object
-        side : {'left', 'right'}
-        kind : {'loc', 'getitem'} or None
-
-        Returns
-        -------
-        label : object
-        """
-        assert kind in ["loc", "getitem", None, lib.no_default]
-        self._deprecated_arg(kind, "kind", "_maybe_cast_slice_bound")
-
-        if isinstance(label, str):
-            try:
-                parsed, reso = self._parse_with_reso(label)
-            except ValueError as err:
-                # e.g. 'unit abbreviation w/o a number'
-                raise self._invalid_indexer("slice", label) from err
-
-            lower, upper = self._parsed_string_to_bounds(reso, parsed)
-            return lower if side == "left" else upper
-        elif not isinstance(label, self._data._recognized_scalars):
-            raise self._invalid_indexer("slice", label)
-
-        return label
-
     def _parse_with_reso(self, label: str):
         # the "with_reso" is a no-op for TimedeltaIndex
         parsed = Timedelta(label)
